ADEN-11320 remove nativo.utils
